Create Topic 3_non synchronized A549 cells normalized with mean value method,  reduced and reproducibility.Rmd
I reduced the data by uniting the replicates. Afterwards I scaled them to 100 again.

Also I changed Tines code for the reproducibility concerning variables and I implemented the as.numeric function.

Also, I had an idea for plotting line graphs which does function, therefore I wanted to push it, but nevertheless it is not nearly perfect and I think, we should search and test other options.